+++ src/aging/providers/main.py
@@ -26,6 +26,8 @@ from ..commands import (
     JumpToMenu,
     SaveEntrySource,
     SaveEntryText,
+    SearchEntry,
+    SearchEntryNextFind,
     SearchForGuide,
     SeeAlso,
     ToggleClassicView,
@@ -59,8 +61,10 @@ class MainCommands(CommandsProvider):
         yield Quit()
         yield from self.maybe(SaveEntrySource)
         yield from self.maybe(SaveEntryText)
-        yield from self.maybe(SeeAlso)
+        yield from self.maybe(SearchEntry)
+        yield from self.maybe(SearchEntryNextFind)
         yield from self.maybe(SearchForGuide)
+        yield from self.maybe(SeeAlso)
         yield ToggleClassicView()
         yield ToggleGuides()